runtime/hermes/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.1
google-generativeai==0.8.4
httpx==0.28.1
pydantic==2.10.4
apscheduler==3.11.0
